requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
vosk==0.3.45
websockets==15.0.1
soundfile==0.13.1
sentence-transformers==5.1.0
scikit-learn==1.7.1
gTTS==2.5.4
python-dotenv==1.1.1
python-multipart==0.0.20
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9